Add Ten Clients More Consumers By DateTime Range

+++ src/main/java/edu/utn/udee/Udee/service/backoffice/AddressService.java
@@ -3,30 +3,40 @@ package edu.utn.udee.Udee.service.backoffice;
 import edu.utn.udee.Udee.domain.Address;
 import edu.utn.udee.Udee.exceptions.AddressExistsException;
 import edu.utn.udee.Udee.exceptions.AddressNotExistsException;
+import edu.utn.udee.Udee.exceptions.ClientNotExistsException;
+import edu.utn.udee.Udee.exceptions.RateNotExistsException;
 import edu.utn.udee.Udee.repository.AddressRepository;
+import edu.utn.udee.Udee.repository.ClientRepository;
+import edu.utn.udee.Udee.repository.RateRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-
 @Service
 public class AddressService {
 
     private final AddressRepository addressRepository;
+    private final ClientRepository clientRepository;
+    private final RateRepository rateRepository;
 
     @Autowired
-    public AddressService(AddressRepository addressRepository){
+    public AddressService(AddressRepository addressRepository, ClientRepository clientRepository, RateRepository rateRepository){
         this.addressRepository = addressRepository;
+        this.clientRepository = clientRepository;
+        this.rateRepository = rateRepository;
     }
 
-    public Address addAddress(Address address) throws AddressExistsException {
+    public Address addAddress(Address address)
+            throws AddressExistsException, ClientNotExistsException, RateNotExistsException {
         if(!addressRepository.existsByAddress(address.getAddress())){
-            return addressRepository.save(address);
-        } else {
-            throw new AddressExistsException();
-        }
+            if(clientRepository.existsById(address.getClient().getId())) {
+                if (rateRepository.existsById(address.getRate().getId()))
+                    return addressRepository.save(address);
+                else throw new RateNotExistsException();
+            }
+            else throw new ClientNotExistsException();
+        } else throw new AddressExistsException();
     }
 
     public Page<Address> allAddress(Pageable pageable){
@@ -34,7 +44,9 @@ public class AddressService {
     }
 
     public void deleteAddressById(Integer id) throws AddressNotExistsException {
-        addressRepository.deleteById(id);
+        if (addressRepository.existsById(id))
+            addressRepository.deleteById(id);
+        else throw new AddressNotExistsException();
     }
 
     public Address editAddress(Address address, Integer id) throws AddressNotExistsException {
@@ -56,13 +68,13 @@ public class AddressService {
         return addressRepository.findById(id).orElseThrow(AddressNotExistsException::new);
     }
 
-    public Address findAddressByAddress(String address){
-        return addressRepository.findAddressByAddress(address);
-    }
-
-    public List<Address> findAddressByClientId(Integer id){
-        return addressRepository.findAddressByClientId(id);
-    }
+//    public Address findAddressByAddress(String address){
+//        return addressRepository.findAddressByAddress(address);
+//    }
+//
+//    public List<Address> findAddressByClientId(Integer id){
+//        return addressRepository.findAddressByClientId(id);
+//    }
 
     /*public Address editAddress(Address address) throws AddressNotExistsException {
         if(addressRepository.)
